refactor: resolve 'require cycle' warning
